Prevent publish with uncommited changes

+++ publish.sh
@@ -2,6 +2,16 @@
 
 set -e;
 
+if ! git diff-files --quiet; then
+    echo "Can not publish with unstaged uncommited changes";
+    exit 1;
+fi;
+
+if ! git diff-index --quiet --cached HEAD; then
+    echo "Can not publish with staged uncommited changes";
+    exit 1;
+fi;
+
 rm -rf node_modules/cross-domain-safe-weakmap node_modules/sync-browser-mocks
 npm install cross-domain-safe-weakmap sync-browser-mocks
